+++ twotower/__init__.py
@@ -2,7 +2,8 @@
 
 from .config import TwoTowerConfig
 from .core import TwoTower
+from .features import FeatureConfig, MultiFeatureSpec
 
 __version__ = "0.1.0"
 
-__all__ = ["TwoTower", "TwoTowerConfig", "__version__"]
+__all__ = ["TwoTower", "TwoTowerConfig", "FeatureConfig", "MultiFeatureSpec", "__version__"]